e2e-tests/requirements.txt
bidpom
PyPOM
pytest~=3.0.0
pytest-selenium
pytest-xdist==1.15.0
requests==2.11.1
